apps/api/requirements.txt
# Flask Core
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0

# Database
# Use newer SQLAlchemy compatible with Python 3.13
SQLAlchemy==2.0.34
# Use SQLite locally; omit psycopg2-binary
# psycopg2-binary==2.9.7
# Align Alembic with latest compatible
alembic==1.17.0

# Authentication & Security
bcrypt==4.1.1
PyJWT==2.8.0
cryptography==42.0.5

# File Handling
Werkzeug==3.0.1
Pillow==11.0.0

# PDF Generation
reportlab==4.0.7
openpyxl==3.1.5

# QR Code Generation
qrcode[pil]==7.4.2

# Email
secure-smtplib==0.1.1

# Validation
email-validator==2.1.0

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2

# Production Server
gunicorn==21.2.0

# Development
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
black==23.12.0
flake8==6.1.0

